upgrade models, remove mailer model
